Visualizzazione prenotazioni e fix grafici
