+++ static/scriptInsertion.js
@@ -39,8 +39,8 @@ async function loadInfo()
             console.log("data", data)
 
             $("h5:eq(0)").text(data.name)
-            $("h5:eq(1)").text(data.datetimeStart)
-            $("h5:eq(2)").text(data.datetimeEnd)
+            $("h5:eq(1)").text((new Date(data.datetimeStart)).toLocaleString("it-IT"))
+            $("h5:eq(2)").text((new Date(data.datetimeEnd)).toLocaleString("it-IT"))
             $("h5:eq(3)").text(data.minInterval + " minuti")
             $("h5:eq(4)").text(data.priceHourly + " EUR/hr.")
             $("h5:eq(5)").text(data.priceDaily + " EUR/gg")
@@ -75,7 +75,9 @@ async function loadInfo()
                 tmpHTML = $(".prenotazList").clone()
                 tmpHTML.removeAttr("hidden")
                 tmpHTML.removeClass("prenotazList")
-                $(tmpHTML).html("da: " + data.reservations[i].datetimeStart + "  a: " + data.reservations[i].datetimeEnd + "da parte di: "+data.reservations[i].client._id)
+                data.reservations[i].datetimeStart = (new Date(data.reservations[i].datetimeStart)).toLocaleString("it-IT")
+                data.reservations[i].datetimeEnd = (new Date(data.reservations[i].datetimeEnd)).toLocaleString("it-IT")
+                $(tmpHTML).html("Da: " + data.reservations[i].datetimeStart + "<br>A: " + data.reservations[i].datetimeEnd + "<br>Da parte di: "+data.reservations[i].client.username)
                 container.append(tmpHTML)
             }
             container.append("</ul>")
